Feature SALES-4: Implements aggregate getter endpoints for book detail (#51)
 Summary                                                                                         
                                                                                                     
  - Implements aggregate financial summary for the Book Detail view (req 2.2.2), providing         
  server-computed totals for publisher revenue, unpaid/paid/total author royalty, and total units    
  sold                                                                                           
  - Introduces BookDetailResponse DTO to carry book metadata + financials in a single request,     
  replacing the previous approach of computing totals on the frontend                              
  - Merges /bookdetail/{id} into /book/{id} so the detail endpoint returns financials inline      
  - Sunsets the redundant PagedBookResponse DTO in favor of the generic PagedResponse<T>
  - Resolves merge conflicts with staging (ISBN lookup feature)

  Test plan

  - Verify Book Detail page displays correct financial summary values
  - Confirm GET /api/books/{id} returns BookDetailResponse with all financial fields
  - Verify Book List view still works with PagedResponse<BookResponse>
  - Run existing backend tests (BookController, SaleService)
  - Confirm ISBN lookup endpoint still functions after merge resolution

  Resolves #20, #49, #67

+++ backend/src/main/java/edu/duke/bookpublishing/sales/SaleRepository.java
@@ -25,7 +25,7 @@ public interface SaleRepository extends JpaRepository<Sale, Long>, JpaSpecificat
             from Sale s
             where s.book.id = :bookId
             """)
-  Integer totalUnitsSoldByBook(Long bookId);
+  Long totalUnitsSoldByBook(Long bookId);
 
   // ---- Requirement 2.2 Book Detail ----
   // 1) Total Publisher revenue for a book
